requirements.txt
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-community>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.0